// src/app/Model/PokemonInfo.ts
export class PokemonInfo{
    No:number;
    Name:string;
    type1:string;
    type2:string;
    Ability: string[];//特性
    Gender: string;//性別
    Color: string;
    Height: number;
    Weight: number;
    Egg: string;
    EggCount: number;
    TotalExp: number;
    HP:number;
    Atk:number;
    Def:number;
    SpAtk:number;
    SpDef:number;
    Spd:number;
}